[easy] Move tests that are not full pipeline tests into their respective packages
