+++ starfish/intensity_table/test/test_intensity_table_coords.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from starfish import IntensityTable
 from starfish.intensity_table import intensity_table_coordinates
-from starfish.test import test_utils
+from starfish.test import factories
 from starfish.types import Axes, Coordinates, Features, PhysicalCoordinateTypes
 
 NUMBER_SPOTS = 10
@@ -30,8 +30,8 @@ def test_tranfering_physical_coords_to_intensity_table():
                                    (PhysicalCoordinateTypes.Z_MIN, 1),
                                    (PhysicalCoordinateTypes.Z_MAX, 3)])
 
-    stack = test_utils.imagestack_with_coords_factory(stack_shape, physical_coords)
-    codebook = test_utils.codebook_array_factory()
+    stack = factories.imagestack_with_coords_factory(stack_shape, physical_coords)
+    codebook = factories.codebook_array_factory()
 
     intensities = IntensityTable.synthetic_intensities(
         codebook,
@@ -81,8 +81,8 @@ def test_tranfering_physical_coords_to_expression_matrix():
                                    (PhysicalCoordinateTypes.Z_MIN, 1),
                                    (PhysicalCoordinateTypes.Z_MAX, 3)])
 
-    stack = test_utils.imagestack_with_coords_factory(stack_shape, physical_coords)
-    codebook = test_utils.codebook_array_factory()
+    stack = factories.imagestack_with_coords_factory(stack_shape, physical_coords)
+    codebook = factories.codebook_array_factory()
 
     intensities = IntensityTable.synthetic_intensities(
         codebook,
